feat: wake add multiple items
