+++ wake/actions/cart/addItems.ts
@@ -1,12 +1,22 @@
 import { HttpError } from "../../../utils/http.ts";
 import { AppContext } from "../../mod.ts";
-import AddItems from "./addItems.ts";
+import { getCartCookie, setCartCookie } from "../../utils/cart.ts";
+import { AddItemToCart } from "../../utils/graphql/queries.ts";
+import {
+  AddItemToCartMutation,
+  AddItemToCartMutationVariables,
+  CheckoutFragment,
+} from "../../utils/graphql/storefront.graphql.gen.ts";
 
-export interface Props {
+export interface CartItem {
   productVariantId: number;
   quantity: number;
   customization: { customizationId: number; value: string }[];
   subscription: { subscriptionGroupId: number; recurringTypeId: number };
+}
+
+export interface Props {
+  products: CartItem[];
 }
 
 const action = async (
@@ -14,7 +24,25 @@ const action = async (
   req: Request,
   ctx: AppContext,
 ): Promise<Partial<CheckoutFragment>> => {
-  return AddItems({ products: props }, req, ctx);
+  const { storefront } = ctx;
+  const cartId = getCartCookie(req.headers);
+
+  if (!cartId) {
+    throw new HttpError(400, "Missing cart cookie");
+  }
+
+  const data = await storefront.query<
+    AddItemToCartMutation,
+    AddItemToCartMutationVariables
+  >({
+    variables: { input: { id: cartId, products: props.products } },
+    ...AddItemToCart,
+  });
+
+  const checkoutId = data.checkout?.checkoutId;
+  setCartCookie(ctx.response.headers, checkoutId);
+
+  return data.checkout ?? {};
 };
 
 export default action;
